Store amounts at their full u64 width
Amounts are u64 on the wire, and int8 held only the lower half of that
range, so a peer could name an amount the mint had to refuse rather than
store. They now go to numeric(20, 0) on Postgres and to zero padded
decimal text on SQLite, whose only integer is signed and which promotes
an overflow to a float instead of failing. Padding keeps lexicographic
order the same as numeric order, so nothing has to cast to sort or
compare.

The Rust type carries the decision: a bound Amount becomes its own value
variant that survives to the driver, because SQLite cannot tell at bind
time which column a bare u64 is headed for. Everything that is not an
amount stays int8.

The two totals computed in SQL go through u64_add and u64_sum, named the
same on both backends so one statement works everywhere. On SQLite they
are registered functions that refuse an overflow; on Postgres they alias
arithmetic the server already does exactly, and a check constraint on
keyset_amounts refuses a running total past u64::MAX.

The branch's separate widening migrations are folded in, since taking a
column to int8 and then to numeric rewrote the same table twice.

+++ crates/cdk-sql-common/src/mint/auth/mod.rs
@@ -185,7 +185,7 @@ where
                    "#,
             )?
             .bind("blinded_message", message.to_bytes().to_vec())
-            .bind("amount", u64::from(signature.amount))
+            .bind("amount", signature.amount)
             .bind("keyset_id", signature.keyset_id.to_string())
             .bind("c", signature.c.to_bytes().to_vec())
             .execute(&self.inner)
